Remove zero_hw_sentinel and weights tracked by statistics from the state_dict. Refactor proxies to inherit from a common abstract class.
